+++ integrations/authorization/v0/implementation.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -51,7 +51,7 @@ func (i *implementation) Name() string {
 	return pbAuthorizationV0.Name
 }
 
-func (i *implementation) Health() svc.HealthState {
+func (i *implementation) Health(ctx context.Context) svc.HealthState {
 	return svc.Healthy
 }