+++ integrations/authorization/v1/impl.go
@@ -1,0 +1,93 @@
+//
+// DISCLAIMER
+//
+// Copyright 2025-2026 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package v1
+
+import (
+	"context"
+
+	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
+	"google.golang.org/grpc"
+
+	pbAuthorizationV1 "github.com/arangodb/kube-arangodb/integrations/authorization/v1/definition"
+	pbImplAuthorizationV1Shared "github.com/arangodb/kube-arangodb/integrations/authorization/v1/shared"
+	"github.com/arangodb/kube-arangodb/pkg/util/svc"
+)
+
+func New(cfg Configuration) (svc.Handler, error) {
+	if err := cfg.Validate(); err != nil {
+		return nil, err
+	}
+
+	plugin, err := cfg.Plugin()
+	if err != nil {
+		return nil, err
+	}
+
+	return newInternal(cfg, plugin), nil
+}
+
+func newInternal(cfg Configuration, plugin pbImplAuthorizationV1Shared.Plugin) *implementation {
+	obj := &implementation{
+		cfg:    cfg,
+		plugin: plugin,
+	}
+
+	return obj
+}
+
+var _ pbAuthorizationV1.AuthorizationV1Server = &implementation{}
+var _ svc.Handler = &implementation{}
+
+type implementation struct {
+	pbAuthorizationV1.UnimplementedAuthorizationV1Server
+
+	cfg Configuration
+
+	plugin pbImplAuthorizationV1Shared.Plugin
+}
+
+func (i *implementation) Name() string {
+	return pbAuthorizationV1.Name
+}
+
+func (i *implementation) Health(ctx context.Context) svc.HealthState {
+	if err := i.plugin.Ready(ctx); err != nil {
+		logger.Err(err).Warn("Service is not ready")
+		return svc.Degraded
+	}
+	return svc.Healthy
+}
+
+func (i *implementation) Register(registrar *grpc.Server) {
+	pbAuthorizationV1.RegisterAuthorizationV1Server(registrar, i)
+}
+
+func (i *implementation) Gateway(ctx context.Context, mux *runtime.ServeMux) error {
+	return nil
+}
+
+func (i *implementation) Background(ctx context.Context) {
+	i.plugin.Background(ctx)
+}
+
+func (i *implementation) Evaluate(ctx context.Context, request *pbAuthorizationV1.AuthorizationV1PermissionRequest) (*pbAuthorizationV1.AuthorizationV1PermissionResponse, error) {
+	return i.plugin.Evaluate(ctx, request)
+}