+++ integrations/pong/v1/impl.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2023-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2023-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -107,7 +107,7 @@ func (i *impl) Name() string {
 	return pbPongV1.Name
 }
 
-func (i *impl) Health() svc.HealthState {
+func (i *impl) Health(ctx context.Context) svc.HealthState {
 	return svc.Healthy
 }