+++ integrations/storage/v2/storage.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -79,7 +79,7 @@ func (i *implementation) Gateway(ctx context.Context, mux *runtime.ServeMux) err
 	return nil
 }
 
-func (i *implementation) Health() svc.HealthState {
+func (i *implementation) Health(ctx context.Context) svc.HealthState {
 	return svc.Healthy
 }