+++ pkg/integrations/authorization_v1.go
@@ -1,0 +1,64 @@
+//
+// DISCLAIMER
+//
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package integrations
+
+import (
+	"context"
+
+	"github.com/spf13/cobra"
+
+	pbImplAuthorizationV1 "github.com/arangodb/kube-arangodb/integrations/authorization/v1"
+	pbAuthorizationV1 "github.com/arangodb/kube-arangodb/integrations/authorization/v1/definition"
+	"github.com/arangodb/kube-arangodb/pkg/util/errors"
+	"github.com/arangodb/kube-arangodb/pkg/util/svc"
+)
+
+func init() {
+	registerer.Register(pbAuthorizationV1.Name, func() Integration {
+		return &authorizationV1{}
+	})
+}
+
+type authorizationV1 struct {
+	config pbImplAuthorizationV1.Configuration
+}
+
+func (a authorizationV1) Name() string {
+	return pbAuthorizationV1.Name
+}
+
+func (a authorizationV1) Description() string {
+	return "Enable AuthorizationV1 Integration Service"
+}
+
+func (a authorizationV1) Register(cmd *cobra.Command, fs FlagEnvHandler) error {
+	return errors.Errors(
+		fs.StringVar((*string)(&a.config.Type), "type", string(pbImplAuthorizationV1.ConfigurationTypeAlways), "Type of the Authorization Integration"),
+	)
+}
+
+func (a authorizationV1) Handler(ctx context.Context, cmd *cobra.Command) (svc.Handler, error) {
+	return pbImplAuthorizationV1.New(a.config)
+}
+
+func (a authorizationV1) Init(ctx context.Context, cmd *cobra.Command) error {
+	return nil
+}